+++ presentation/src/main/java/com/shypolarbear/presentation/ui/common/ImageViewPagerViewHolder.kt
@@ -1,11 +1,11 @@
-package com.shypolarbear.presentation.ui.feed.viewholder
+package com.shypolarbear.presentation.ui.common
 
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.shypolarbear.domain.model.feed.FeedPostImg
 import com.shypolarbear.presentation.databinding.ItemFeedPostImgBinding
 
-class FeedPostImgViewHolder(private val binding: ItemFeedPostImgBinding) : RecyclerView.ViewHolder(binding.root) {
+class ImageViewPagerViewHolder(private val binding: ItemFeedPostImgBinding) : RecyclerView.ViewHolder(binding.root) {
 
     fun bind(img: FeedPostImg) {
         Glide.with(itemView)
